topology_task: Fragment device neighbor TLV
In case there is a large number of neigbors, the CMDU might need to be
fragmented. Since the fragmentation is possible only on TLV boudaries,
TLV larger than MTU size are impossible to fragment.

To overcome this issue - a new TLV need to be when we run out of space.
This commit adds the fragmentaion logic to the topology_task neighbor
device TLV.

Signed-off-by: Semyon Barenboym <semyon.barenboym@devalore.com>