transport: END_OF_MESSAGE TLV expected only in last fragment
IEEE1905 standard expects END_OF_MESSAGE TLV to be added only on the
last fragment of a fragmented CMDU.
The code today was adding the EOM TLV to each fragment.
This commit removes the EOM TLV from fragments on fragmentation side.
On defragmentation side it removes the considation of EOM TLV that is
added to each fragment.

Signed-off-by: Semyon Barenboym <semyon.barenboym@devalore.com>