transport: remove interface on error
https://jira.prplfoundation.org/browse/PPM-1204

The transport process must deal with network interfaces that belong to
the bridge, that have an index and a MAC address and that are up and
running. If whatever of these conditions is not met, then the interface
must not be used.

Add a new method to remove a network interface and call it whenever any
of the previous conditions is not met.

Signed-off-by: mmazzapater <mmazzapater@gmail.com>