controller: db: Fix duplicate STA instance
In current implementation function for updating STA data
same method called where present function for adding
new instances of STA to the Data Model without checking
that instance exist on Data Model or not. This cause the
problem with multiple STA instances on the Data Model
for same device.

Add verification that STA is unique before adding
new instance to the Data Model.

Signed-off-by: Vladyslav Tupikin <v.tupikin@inango-systems.com>