agent: channel_scan_task: Set timestamp for first neighbor
When receiving channel scan results from the monitor, we save the
results in a map based on the requested timestamp.
When the map is already populated with old results, we first clear the
vector to make it ready for the new scan results.
It was seen that when we recieve the first ever result for a channel, we
missed setting the timestamp value to the current one, causing the next
result to flush the first result.

This commit adds the setting of the first ever timestamp.
PPM-1051

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>
Signed-off-by: Semyon Barenboym <semyon.barenboym@devalore.com>