tools: deploy_firmware.py: expect the output for a successful write
In addition to timing out because it repeatedly tries to download the
image from the tftp server, sometimes the upgrade fails right away in
this way:

TFTP from server 192.168.250.199; our IP address is 192.168.250.180
Filename 'AX3000_1600_ETH_11AXUCI-squashfs-fullimage.img'.
Load address: 0x80800000
Loading: *
ARP Retry count exceeded; starting again
Image not found!!
GRX500 # reset

In this case, there is no pexpect timeout since we do get the prompt
back. To handle it, add expect lines for a successful image write.

There is no line that specifically says that the overall upgrade was
successful, so instead check if at least one UBI volume was created,
and if we do get the "Writing to NAND" line at least once.

While at it, add a print statement to make it clear when the upgrade
is done from pexpect's point of view.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>