transport: Small fragment padding warning
When fragmenting a message it is expected that only the last fragment
will have an END OF MESSAGE TLV.
However, when sending a TLV smaller than then the minimum ethernet
frame, a padding is added to the frame.
This padding might cause us having unwanted EOM TLV.

This commit adds a constant value of minimum fragment length and a
warning when sending small fragment.

Signed-off-by: Semyon Barenboym <semyon.barenboym@devalore.com>