agent: backhaul_manager: delay enabling transport
https://jira.prplfoundation.org/browse/PPM-959

The certification testbed can contain running controllers from previous
tests.

prplMesh is started, then the test begins and every device is sent
dev_reset_default, which should make them stop acting as controllers.

Before the dev_reset_default however, the local prplMesh agent that
goes with the prplMesh controller already looks for a controller. As a
consequence it's possible that the local agent will be onboarded to
another controller.

When prplMesh receives dev_reset_default, the local agent is not reset
and stays connected to the (now non-existent) controller. From prplMesh
controller's point of view, the local agent didn't connect.

To fix this problem, in certification mode, if prplMesh is configured
with local controller, do not enable the transport process until agent
has connected to controller.

Signed-off-by: mmazzapater <mmazzapater@gmail.com>