agent: move iw copyright notice
prplMesh uses some code borrowed from `iw`. This code was initially
placed in backhaul_manager.cpp but at some point in time it was moved
to capability_reporting_task.cpp. However, the copyright notice was not
moved along.

Move the copyright notice that must accompany the `iw` code to the file
where it is actually used.

Signed-off-by: mmazzapater <mmazzapater@gmail.com>