bwl: nl80211: read preferred channels
https://jira.prplfoundation.org/browse/PPM-1221

Currently, in a Turris-Omnia device, method `read_preferred_channels()`
is filling the list of preferred channels with both 5Ghz and 2.4Ghz
channels. On the contrary, method `refresh_radio_info()` is filling
the list of supported channels with the 5Ghz channels only.

This inconsistency (2.4Ghz channels are preferred but not supported)
leads to a crash in AP manager.

To fix the problem and avoid the crash, fill in the list of preferred
channels with the supported channels only.

Signed-off-by: mmazzapater <mmazzapater@gmail.com>