tlvf: WSC::m1: copy the buffer to allow parsing multiple times
The TLVF parse functions modify the underlying buffers because the data
is swapped. Because of this, it is not possible to parse the same buffer
twice.

For CMDUs, parsing twice is avoided by just parsing it once immediately
after receiving it, and using getClass() only after that.

For WSC, this is not so easy. We could do the parsing in parseNextTlv(),
but then we have nowhere to save the parsed M1 or M2 (there is no
WSC::m1 or WSC::m2 data member in the tlvWsc, and it can't be added
since the tlvf code generator doesn't support such a thing).

Therefore, as an alternative, make a copy of the buffer when parsing the
WSC message. To make sure the buffer is freed on deletion, the m1 class
is subclassed to WSC::M1Buff which has the copy as a std::vector member.
WSC::m1::parse() returns such an object instead of a plain m1.

This way, WSC::m1::parse() can be called multiple times on the same
tlvWsc.

For now, only do it for m1 because that's the only one where we need
this feature.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>