bwl: dwpal: read preferred channels
https://jira.prplfoundation.org/browse/PPM-1221

Method `read_preferred_channels()` is filling in the list of preferred
channels with all channels from all bands. However, the list of
supported channels contains the channels of the first band found. If we
would ever use a dual-band device with DWPAL, the list of preferred
channels would contain more channels than the list of supported
channels.

To avoid this problem, implement the DWPAL version of
`read_preferred_channels()` the same way as the NL80211 version, by
copying  the list of supported channels into the list of preferred
channels. Supported channels are read right after opening the control
interface, on the GetRadioInfo FSM state.

This implementation is also more efficient, as we do not have to query
for some information that we have read already.

Signed-off-by: mmazzapater <mmazzapater@gmail.com>