bwl: read radio info from ap_wlan_hal_*
https://jira.prplfoundation.org/browse/PPM-366

This is a preparatory commit.

Method `refresh_radio_info()` in base_wlan_hal_* is filling in the list
of supported channels. To do this properly in a dual-band radio, we
need to know which band is being used. To know the band, we'll use a
parameter in the hostapd.conf file. To read this parameter we need
functions that are defined in derived class ap_wlan_hal_*.

So override and extend `refresh_radio_info()` in ap_wlan_hal_*. Move the
code that fills in supported channels, as is, from base to derived class
method. In a follow-up commit this code will be modified to compute used
band instead of choosing a band arbitrarily.

Signed-off-by: mmazzapater <mmazzapater@gmail.com>