bwl: nl80211: use band set in hostapd.conf
https://jira.prplfoundation.org/browse/PPM-366

When radio information is read for a dual-band radio using the
NL80211_CMD_GET_WIPHY command, two bands are reported. However, no clue
about which one is being actually used is included in the response.

Read parameter `hw_mode` in the hostapd configuration file to know which
band in a dual-band device is being used and then fill in the radio
information accordingly.

Most of the radio information never changes so it could be read only
once. However, attributes like the DFS state do change and, to read the
latest value, we need to refresh data on each call. So remove the
condition that allowed reading data only the first time this method was
called.

This commit modifies method `refresh_radio_info()` in class
`ap_wlan_hal_nl80211`. Next commit will use this same code in class
`ap_wlan_hal_dwpal`.

Signed-off-by: mmazzapater <mmazzapater@gmail.com>