controller: db: add channel scan is pending flag
Currently when a single scan is requested, but has not yet been
triggered. The "Scan in progress" flag is set to 'false'.
This causes the "get scan results" to return an empty results list.

This is incorrect, in the case where a single scan was requested it
should return "scan in progress".

To resolve this, add a flag to indicate that a scan is in pending, and
when we check for the "scan in progress status" consider the resource
as busy.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>