base_hal_dwpal: update vap info on ssid change
On external SSID configuration the prplmesh doesn't update the SSID
configuration for the VAP. This is because in the refresh_vap_info()
that is called by the ap-manager, change in SSID is not checked and the
VAP info in the available_vaps container is not updated.

Then, the ap-manager propagates the old SSID to the controller which
updates in the controller DB the old SSID.
The above may cause steering operations to fail since source and target
VAPs are compared to have the same SSID. It will also cause wrong
bml-conn-map output.

Added check for change in SSID and if SSID changed, the VAP info is
overridden with the new information.

PPM-1243.

Signed-off-by: Adam Dov <adamx.dov@intel.com>