controller:nbapi:association_event: modify association event
According to TR181 NBAPI AssociationEventData element should be a list of objects.
In this list each object should describes an event generated when a STA associates to a BSS.
Currently we have AssociationEventData singleton object with data only about the last client
associated to BSS. Also, the names of object is incorrect we have 'AssocData'
instead of 'AssociationEventData' and timestamp is missing.

Change AssociationEventData from singleton object to template object.
Add instances of AssociationEventData each time when STA associate to BSS.
Add timestamp with time when this event was collected.
Saved indexes of AssociationEventData in map to be able to get object in future.
As Ambiorix get_object_index method dosen't work with list of objects with the same key.
There are multiple AssociationEvents with the same key (client MAC).

PPM-1240
--
Signed-off-by: Krystyna Oliinyk <k.oliinyk@inango-systems.com>