controller:nbapi:disassociation_event: modify nbapi disassociation event
According to TR181 NBAPI DisassociationEventData element should be a list of objects.
In this list each object should describes an event generated when a STA disassociates
from a BSS. Currently our DisassociationEventData is a singleton object with
data only about the last client which leave BSS.
Also, the names of object is incorrect we have 'DisassocData'
instead of 'DisassociationEventData' and timestamp is missing.

Change DisassociationEventData from singleton object to template object.
Add instances of DisassociationEventData when STA disassociates from a BSS.
Add timestamp with time when this event was collected.

PPM-1240
--
Signed-off-by: Krystyna Oliinyk <k.oliinyk@inango-systems.com>