base_wlan_hal_dwpal: add API for blocking NL reads
To change the "is_zwdfs_supported()" implementation, which currently
relys on number of VAPs on that radio, to read the information from the
NL directly, and to make the read blocking - it is required to use the
new API of dwpal_nl80211_cmd_send().

Added API that utilizes the dwpal_nl80211_cmd_send() which performs
blocking NL reads and parsing the reply using a provided callback.

This API is generaic and can be used for any NL command.

The API will be used in the following commit to read WIPHY information
and determine if an interface supports ZWDFS or not.

Signed-off-by: Adam Dov <adamx.dov@intel.com>