ap_wlan_hal_dwpal: use NL command to check for ZWDFS support
Added static function to check if an interface supports ZWDFS by reading
the WIPHY info and checking for the existence of the
NL80211_ATTR_WIPHY_DFS_ANTENNA attribute.

If exists reutrn true, otherwise return false.

Call the base class API introduced in previous commit with the callback
that checks for ZWDFS support.

Signed-off-by: Adam Dov <adamx.dov@intel.com>