ci: reboot the host if the Marvell STA driver failed to load
From time to time, the Marvell station driver fails to load.

The Marvell station being a PCI device, we already have a script to
try to reset it before loading the driver. It works in some cases, but
as can be seen in [1] it doesn't always work.

On the other hand, we know that rebooting the host completely fixes
the issue. Since the Marvell STA runs on the device running the Gitlab
Runner, the best we can do for now is to reboot the host if the driver
failed to load at the end of a job.

Add a script to check the system logs for the line indicating the
driver failed to load. If it's found, reboot the host.

Closes PPM-1248

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/1154105595

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>