ci: allow manual run of certification tests
There are already several ways to "manually" run certification tests:
- individual tests can be triggered from the pipeline page - but then
  it's not possible to run several tests or to choose the runner;
- a set of tests can be run from a schedule or from a pipeline trigger -
  but that's not possible for an existing pipeline, a new pipeline
  always has to be created for it.

Allow to run a set of tests also after a manual trigger. Since there's a
separate job for the Essensium and the ISR testbeds, it's easy to choose
which testbed to run on. It's also possible to choose any device. It's a
bit more clunky to run that way, because you need to know exactly which
variables to set, but it's anyway more practical than triggering an
entirely new pipeline.

There's still an important limitation, though: it's not possible to
change the variables (target platform, set of tests) after the job has
been triggered once.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>