treewide: do not print error message after getopt
According to its man page, by default, getopt() prints an error message
on standard error, places the erroneous option character in optopt, and
returns '?' as the function result.

Current code is printing an additional error message to standard output
in that case. However, the invalid option is printed as an integer, not
as a character, which renders the message hard to use for debugging
purposes.

Since getopt() is already printing an error message, remove any
redundant messages.

Signed-off-by: mmazzapater <mmazzapater@gmail.com>