TLVF: Fix element allocation in a simple list
When allocating a new element(s) in a TLV, it is expected to copy the
rest of the buffer forward to avoid data loss of already assigned
fields.
While testing, it was observed that there is a data override when
allocating each element by itself instead of a bulk allocation.
The issue was due to incorrect calculation of the src pointer. Instead
of jumping over already allocated elements in the simple list - the src
ptr was taken to the start of the list.
This commit fixes the pointer calculation.

Signed-off-by: Semyon Barenboym <simonbaren@gmail.com>