tools: sim: if link is not in downstream bridge it is broken
We can't consider a link to be part of the backhaul path if it is not in
the bridge. Up to now, we only check the upstream bridge (that's what we
iterate over) not the downstream bridge.

Add a test for the two situations that a link is only in the downstream
bridge or only in the upstream bridge.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>