tools: sim: Link: add 'active' state
We will want to simulate links that dynamically become inactive and
active again. Add an active attribute to the Link class to model this.

If a link is inactive, it is not taken into account for the backhaul
path. Update add_paths_to_neighbors accordingly, and update
test_connectivity_chain_broken to use the active attribute to "destroy"
a link rather than manipulating the bridge. Note that the _upstream and
_downstream tests still make sure that manipulating the bridge is also
tested.

If a link is inactive, messages can't be forwarded over it. Update
forward_message accordingly. Add a new test_message_unicast_broken_link
to validate this.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>