agent: Move channel_scan_policy handling to backhaul manager
The channel scan policy TLV is sent to configure a whole agent policy
and not a specific radio. Thus, it is better to handle it in a more
common place than it is handled today, in the son_slave.
This commit moves the channel_scan_policy in the DB from radio context
to a common context and migrate the handling of the channel scan policy
to backhaul manager context - link_metrics_collection_task.

Signed-off-by: Semyon Barenboym <simonbaren@gmail.com>