bwl: shared: remove thread_local from send_receive_msg
https://jira.prplfoundation.org/browse/PPM-1270

In method `netlink_socket::send_receive_msg()`, we need to set a
callback function to handle the response to a Netlink request. This is
a C function that must delegate the handling of the message to the
`std::function` provided by the calling party as a parameter to the
method.

Passing a lambda with capture is not supported for standard C function
pointers. As a workaround, we were creating a static (but thread local)
wrapper function that was calling the capturing lambda function.

It has been reported that with some compiler, the process crashes right
in the line of code with the `thread_local` specifier and that there is
no such crash if the specifier is removed.

It is not possible to simply remove the usage of the `thread_local`
specifier when declaring the variable because the method is used from
different threads and each thread needs a different copy of the
variable.

So refactor the method around the creation of the C callback function
to handle the response message to avoid using the `thread_local`
specifier. Instead, pass the user's handler function as the `void*`
argument to be passed to the callback. And, inside the C callback, cast
that pointer back to a `std::function` and invoke it. This way the
user's handling function is passed around and there is no need to store
it.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>