agent: Remove duble definition of eCacCompletionStatus
eCacCompletionStatus is defined in the tlvf and on "cac_status_interface.h".
The double definition is redundant, so remove the redefinition on "cac_status_interface.h"
and instead, use the definition from tlvf.

In Addition, add a default value to the CAC status on the tlvf, and remove the CAC status
from the agent_db since it is not being used.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>