client_steering_task: treat no-steering-needed as success
To consider a scenario where client might independently steer to the
target bssid during the steering_task operation or receiving a request
to a target the client is already on, we need to check if the client
current bssid equals to the target bssid and if so break.

While at it, added a data member to determine if a steer operation was
performed and not set the unfriendly flag for that client if no steer
was performed.

Signed-off-by: Adam Dov <adov@maxlinear.com>