tools: sim: test_sim: allow multiple simultaneous expectations
Up to now, the Expect checker ordered the expectations by their time
(`when`) and iterated over them in order.

This is a problem when multiple expectations "arrive" at the same time.
The order in which they will be triggered is not always predictable.

Therefore, change the expectations from the list to a dictionary of
lists. Find the corresponding expectation in that list by matching what
and arg. This way, we still match when, what and arg. I.e., it stays
equivalent to `assert (when, what, arg) == expected` but no longer
relies on the expectations to be in the correct order.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>