bwl: nl80211: remove __thread from send_nl80211_msg
https://jira.prplfoundation.org/browse/PPM-1287

In method `base_wlan_hal_nl80211::send_nl80211_msg()`, we need to set a
callback function to handle the response to a Netlink request. This is
a C function that must delegate the handling of the message to the
`std::function` provided by the calling party as a parameter to the
method.

Passing a lambda with capture is not supported for standard C function
pointers. As a workaround, we were creating a static (but thread local)
wrapper function that was calling the capturing lambda function.

It has been reported that with some compiler, the process crashes right
in the line of code with the `__thread` specifier and that there is no
such crash if the specifier is removed.

It is not possible to simply remove the usage of the `__thread`
specifier when declaring the variable because the method is used from
different threads and each thread needs a different copy of the
variable.

So refactor the method around the creation of the C callback function
to handle the response message to avoid using the `__thread` specifier.
Instead, pass the user's handler function as the `void*` argument to be
passed to the callback. And, inside the C callback, cast that pointer
back to a `std::function` and invoke it. This way the user's handling
function is passed around and there is no need to store it.

This fix is the same that was applied in PPM-1270 to method
`netlink_socket::send_receive_msg()`, which will eventually replace the
deprecated `base_wlan_hal_nl80211::send_nl80211_msg()`.

Signed-off-by: mmazzapater <mmazzapater@gmail.com>