controller.cpp/agent_monitoring_task.cpp: create agent monitoring task
Need to create an agent monitoring task in the controller.

When an agent connects to the controller, an appropriate
procedure has to be followed to maintain it.

Steps in the task:
+ Topology query/response;
+ Policy configuration;
+ AP capability query/response;
+ Start dynamic channel selection task (currently done from M1 handler)

Add agent monitoring task to controller.
Move policy configuration ap capability, starting of dynamic channel selection task,
ap capability query from controller to agent monitoring task.

PPM-992

Signed-off-by: Krystyna Oliinyk <k.oliinyk@inango-systems.com>