agent: replace misuse of std::transform
std::transform native usage is to iterate on a container and
modify each element on the container.
Unfortunately, there are several places which this function is
used to insert elements into a container.
There is no reason to use it like that since using simple ranged
for-loop is much more readable and simple.
Change all occurrences of using std::transform to simple ranged
for-loop.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>