tlvf: Refactor CAC enums
eCacMethod is used both on the CAC capabilities tlv and on the CAC request tlv.
To prevent definition duplication, move the definition exclusively to the
CAC capabilities file, and include it from the CAC request file.

In addition, shorten the enum values name removing the "CAC_METHOD_" prefix,
and add an enum to CAC completion action status, which was missing.

On the next commit, change the code on the agent to work with this commit changes.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>