agent: Remove redundant members on Coordinated CAC Task
Coordinated CAC task saved pointers to TLVs on the cmdu_rx as members of the class.
This is an explicit bug since the tlv holds pointers to the rx_buffer, which will be
overwritten when another message will be received to the task thread.

Replaced the TLV pointers member, with the real information the task needs as a copy
of the data from the TLV.

Also, removed redundant arguments from the function "send_preference_report()".

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>