tools: deploy_firmware: add support for the Axepoint's default bootloader
Since commit a398d2a0423ec59ad95f351302952b3dda3b09b3, we check for
successful writes of the new firmware. It was tested on a NEC device,
but the Axepoint uses a different string which makes it fails like in
[1].

The new string we have to support is "Writing to Nand... done". Do a
case-insensitive match to be able to support both, and check for
either "ok" or "done".

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/1189964074

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>