agent: ap_manager: add CMDU client factory
https://jira.prplfoundation.org/browse/PPM-966

Add CMDU client factory member variable in ap_manager_thread.
Create CMDU client factory in main and pass it to the class constructor.

This is a preparatory commit. Next commits will use provided CMDU
client factory to dynamically create CMDU client instances connected to
the CMDU server that is running in the slave.

Signed-off-by: mmazzapater <mmazzapater@gmail.com>