boardfarm: Search for ClientAssociationEvent
EasyMesh specifies that the Topology Notification message will have One
AL-MAC TLV and zero or one Client Association Event TLV.
Previous commits added the AL-MAC TLV to the message which caused
boardfarm test to fail when checking the Topology Notification message.
The failure was due looking at the TLV at place 0 regardless of the
order of the TLVs in the message.
This commits adds a search for the Client Association Event TLV in the
CMDU.

Signed-off-by: Semyon Barenboym <simonbaren@gmail.com>