agent: ap_manager: refactor FSM method
https://jira.prplfoundation.org/browse/PPM-966

This is a preparatory commit.

The method implementing the FSM in AP manager is called periodically by
the AP manager thread. In a follow-up commit the AP manager will not
be a thread anymore. Instead, it will be the event loop running in the
main thread who will periodically call such FSM method. The FSM method
must return a boolean to tell the event loop if execution must
terminate.

Signed-off-by: mmazzapater <mmazzapater@gmail.com>