controller: db: cast to node_slave when iterating all nodes
There are two loops in DB where all slave nodes are looked for.
Explicitly cast these nodes to `node_slave` to use `hostap`
parameter. (It is a preparation to move this parameter to
`node_slave`)

Since `node_slave` always initalizes its `hostap` parameter
there is no need to check it against `nullptr`.

Situations when a node changes its type is a bug and needs
to be adressed in a different MR.
This commit does not handle this sitation AT ALL.

PPM-1057
-
Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>