tools: get-git-hash.sh: new script to put all hashes in one place
We refer to a lot of different repositories, and we refer to a specific
hash within these repositories.

Currently, these references are spread out all over the tree. This has a
number of problems:
- it's easy to have duplicates;
- when you need to update a hash, it's hard to find where it was
  defined;
- when you need to check out a repository manually (e.g. to do a manual
  build of something that is normally done inside a docker), it's hard
  to find which value to use.

Therefore, create a file that stores all the references to repositories
and hashes.

Since whenever this file is used, we also need to parse it and do a git
clone and git checkout, create a script that does all that. While we're
at it, also add support for checking out in a different directory, and
for updating an already existing checkout.

Since we will use this in Dockerfiles where it needs to be ADDed, put
the specifications and the script in the same file, get-git-hash.sh.
That way, we don't have to ADD two separate files.

Use get-git-hash.sh for the RDK-B build.

Docker ADD requires the file to reside in the same directory as the
Dockerfile, and doesn't allow symlinks that point out of it. Therefore,
build.sh has to make a copy of the file into the Dockerfile directory.
Add this copy to .gitignore.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>