builder: rdk-b: update and add NBAPI/ambiorix support
Update meta-prplmesh to a version that includes NBAPI. Update the build
process to include Ambiorix, as per the README in meta-prplmesh.

For the time being, we need to add "fakeroot" to the docker image for
that. That's because meta-amx sets [1]:

HOSTTOOLS_append = "fakeroot"

And Yocto Reference Manual says

> If a tool specified in the value of HOSTTOOLS is not found on the
> build host, the OpenEmbedded build system produces an error and the
> build is not started.

This is a bug in the Ambiorix Makefiles, because they don't actually use
fakeroot. It will be fixed in some future release.

[1] https://gitlab.com/soft.at.home/buildsystems/yocto/meta-amx/-/blob/master/conf/layer.conf#L12

https://jira.prplfoundation.org/browse/PPM-1316

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>