agent: channel scan task: cache incoming results
To better align with the failure scenarios we can no longer store the
results of the ongoing scan directly in the Agent-DB.
This is because in case of a failure during the ongoing scan all the
cached results are to be considered "partial". and the BD should store
complete and finalized results.

To solve this, first the results need to be cached within the task.
Then once the scan process is done (successfully) the cached results
need to be stored in the DB for later use.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>