agent: channel scan task: set failure status.
If a scan fails because of an Abort notification, a timeout event, or
any other similar failure, each channel's status needs to be updated.
Because we currently have no way to know which channel failed, we set
the status for all the channels.
Later, when retrieving the results for the report, we will check each
channel for their results' status and add them to the report
accordingly.

The following scenarios update the failure status.
- When a scan's state reaches a Timeout event.
- When a CHANNEL_SCAN_ABORTED_NOTIFICATION is received.
- When there is a failure during the trigger_radio_scan function.

And to reiterate, if the scan fails, we update all the scan's
corresponding channels' status because, currently, we have no way to
identify which channel the scan failed.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>