agent: channel scan task: clear remaining pointers
When handling the channel scan, smart pointers are used.
When the entire scan process is done and a report is about to be sent.
Clear the current_scan structure to remove any remaining memory.
Note that if the scan process was successful, the cached results are
stored in the Agent's DB beforehand.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>