backhaul_manager: allocate empty extended link metrics
PPM-1260
After adding the extended STA link metrics, 4.7.7 fails.
The failure Most likely this is because the older wireshark version
reports the extended link metrics with the same field name as the
original link metrics, so the filter returns the num_bssid of the
extended link metrics, which is indeed 0.

This commits adds a sipmle W/A - allocate empty extended link metrics so
the filter returns none 0 value. The real values will be filled
correctly in PPM-1259.

Signed-off-by: Semyon Barenboym <simonbaren@gmail.com>