ci: certification: refactor template in preparation for R2
In a follow-up commit we'll run 'run_test_file.py' for a different
revision (R2). Add the EASYMESH_REVISION to specify it.

Also set the path to the easymesh_cert repository to
"~/easymesh_cert", since the R2 tests run as the gitlab-user
account (without root permissions). The repository has already been
prepared on the runners manually. Unfortunately, it's not possible to
set the path with a variable because of [1].

Because GIT_CLONE_PATH, power_cycle.py and the after_script are
specific to the R1 testbeds, move them to a new R1-only job.

The jobs that extend '.run-certification-tests-R1' use the new Gitlab
"reference" tag ([2]) to re-use the before_script and add more steps
to it. Unfortunately only one level of nesting is supported, so the
"git-clean-reset" part in '.run-certification-tests-R1' has to be
repeated.

Another job for R2 will be added in a follow-up commit.

While at it, also remove the sudo call (it's not needed anymore), an
"echo" leftover, and turn the short options into long options.

[1]: https://gitlab.com/gitlab-org/gitlab-runner/-/issues/1809
[2]: https://docs.gitlab.com/ee/ci/yaml/#reference-tags

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>