agent: channel scan task: reword cached results comment
As part of PPM-1134 an update was added to cache the incoming results
locally within the task. and only when the the scan is done would the
AgentDB be updated.

Since the cached results might not contain any results from some of the
requested channels we need to iterate over all the requested channels,
and update the DB accordingly.

If a requested channel is not present in the cached results we can
assume the channel did not contain results. Therefore we need to clear
the stored results from the DB to reflect the most recent results.
Also, if a request channel is present in te cached results we need to
override any existing stored results.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>