bml: add API to configure logger by external application
The BML library can be linked by external applications. An example of
such application is the beerocks_cli - which instatiates beerocks::logging
object to configure logging preferences.

However, for C application which cannot use the beerocks::logging class.
Intead, we need to provide a simple interface to acheive the same
logging configuration capabilies by providing an API that manages the
configuration internally.

Added an API that applies the logger settings according to provided
configuration file.

Signed-off-by: Adam Dov <adov@maxlinear.com>