boardfarm: _device_wait_for_log: use ssh and cat for check_log
Each call of `check_log` on a real device maps to
an invocation of `_device_wait_for_log`.
This method is notorious for producing an enormous
amount of output in the boardfarm console.

The reason is that it uses `tail -f` that dumps
log file(s) until the desired line is found.

Changes introduces with this commit:
* Use a separate SSH connection to not have all the
  log file content in the main boardfarm console.
* Use `cat` in a loop instead of `tail -f`. It is less
  performant but easier to implement. It is also more
  consistent with `_docker_wait_for_log`.
* Return `match.groups` instead of `match` object
  as `_docker_wait_for_log` does.

PPM-1283

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>