boardfarm: command: avoid echoing supplied command
`command` method is used to get exernal command output
in a device console.

It is implemented by capturing everything between sending
a command and receiving shell prompt.
Unfortunately, device shell echoes supplied command
and `stty` is absent to fix it (at least on RAX40).

Moreover, if the command is too long to not fit into
console width, it is wrapped. '\r' symbols are inserted.

For the sake of simplicity, execute every command in
a separate shell (SSH connection).

Convert `\r\n` to `\n` in `get_ssid`. `\n` was converted
to `\r\n` by boardfarm console. After switching to SSH
it needs to be removed.

PPM-1293

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>