agent: Add configuration protection
The traffic separation policy can be set on the autoconfiguration
message which is sent separately to each radio.
There could be a scenario in which a radio gets some configuration
TLV, and then the other radio gets not configuration TLV which will
cause to reset the configuration.
Added protection that creates a timezone of 5 seconds in which the
configuration cannot be overridden if configuration TLV is not found.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>