bcl: Add set_vlan_packet_filter() function of network_utils
Configuring VLAN aware bridge is not in full compliance with the Multi-AP
standard traffic separation requirements:
1. Untagged port allows packets with VLAN tag of the port to pass through
but the standard says that untagged port should drop any tagged traffic.
2. VLAN aware bridge which configured to protocol 802.1Q (by default)
does not recognize double-tagged packets with S-Tag (protocol 802.1ad,
aka Q-in-Q). The standard says that when such a packet is received, the
Agent should strip the S-Tag or Drop the packet.

Add a function that adds/removes to the ebtables rules that drop 802.1Q
tagged packets on a given interface name, and drop double-tagged packets.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>