ci: test-on-target: run dctest clean
Recently the test-on-target tests stopped working because they
couldn't access the devices.

After investigating it turned out that the interface used to connect
to the devices was still part of a docker bridge.

This could be due to a docker-compose test that was canceled in the
middle of a run or didn't exit properly.

One way to mitigate this issue is to make sure the docker networks are
always properly removed.

To do that, change dctest.py's clean option:

- Don't remove the images. It's the responsibility of the runners to
free up some space by removing old images after some time.

- After "docker-compose down", also remove all the networks created
for Boardfarm. First list them using their name, then use "docker
network rm" to remove them. Note that unlike "docker-compose down",
this applies to all the networks that are not in use, and not only to
the ones that correspond to the current 'bid'.

Since there is currently no way to guarantee that an after_script is
run (see [1]) run it in a before_script.

[1]: https://gitlab.com/gitlab-org/gitlab-runner/-/issues/4843

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>