controller: db: revert node_stats_info removal
We discovered that when STAs connected, their RSSI field was not
updating correctly and was stuck in its default value of -127.
After examination, we also found that as part of PPM-949 & MR !2201
the flow that updated the STA node in the DB was removed and supposedly
replaced with the Associated STA traffic Stats TLV handling.
But as was seen, the STA's RSSI information was not updated in this new
flow.

To resolve this, reverted the relevant parts of MR !2201.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>