bpl: uci: refactor bpl_cfg_get_wifi_credentials
https://jira.prplfoundation.org/browse/PPM-836

This is a preparatory commit.

In a follow-up commit, a method to read wireless settings from all
fronthaul interfaces is going to be created. This new method will
require part of the code currently in `bpl_cfg_get_wifi_credentials`.

To avoid duplicating code, refactor this common code out of method
`bpl_cfg_get_wifi_credentials`.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>